graph-llm-system/backend/requirements.txt
fastapi
uvicorn[standard]
networkx
psycopg2-binary
google-genai
python-dotenv
